console: move formatTime into internal/util/debuglog
Node exposes formatTime from internal/util/debuglog and its test imports it
from there under --expose-internals. Bun's copy was a closure inside the
ConsoleObject builtin, unreachable by name. It now lives in
internal/util/debuglog, required by createConsoleConstructor the same way that
function already requires internal/validators, and is exposed through
exposedInternals.

Output is unchanged: the algorithm is line-for-line the same and the module is
loaded lazily by the console.Console getter, so global console.timeEnd, which
is native, is untouched. Plain prototype methods replace the captured-$call
style because that is the internal-module convention, per neighboring modules.

Adds test-console-formatTime from Node v26.3.0, verbatim.
