+++ src/js/builtins/ConsoleObject.ts
@@ -155,18 +155,11 @@ export function createConsoleConstructor(console: typeof globalThis.console) {
   const StringPrototypeRepeat = String.prototype.repeat;
   const StringPrototypeSlice = String.prototype.slice;
   const ObjectPrototypeHasOwnProperty = Object.prototype.hasOwnProperty;
-  const StringPrototypePadStart = String.prototype.padStart;
-  const StringPrototypeSplit = String.prototype.split;
-  const NumberPrototypeToFixed = Number.prototype.toFixed;
   const ArrayPrototypeMap = Array.prototype.map;
   const ArrayPrototypeJoin = Array.prototype.join;
   const ArrayPrototypePush = Array.prototype.push;
 
   const kCounts = Symbol("counts");
-
-  const kSecond = 1000;
-  const kMinute = 60 * kSecond;
-  const kHour = 60 * kMinute;
 
   const internalGetStringWidth = $newCppFunction("stringWidth.cpp", "jsFunctionBunStringWidth", 1);
 
@@ -760,39 +753,7 @@ export function createConsoleConstructor(console: typeof globalThis.console) {
     return true;
   }
 
-  function pad(value) {
-    return StringPrototypePadStart.$call(`${value}`, 2, "0");
-  }
-
-  function formatTime(ms) {
-    let hours = 0;
-    let minutes = 0;
-    let seconds: string | number = 0;
-
-    if (ms >= kSecond) {
-      if (ms >= kMinute) {
-        if (ms >= kHour) {
-          hours = Math.floor(ms / kHour);
-          ms = ms % kHour;
-        }
-        minutes = Math.floor(ms / kMinute);
-        ms = ms % kMinute;
-      }
-      seconds = ms / kSecond;
-    }
-
-    if (hours !== 0 || minutes !== 0) {
-      ({ 0: seconds, 1: ms } = (StringPrototypeSplit.$call as any)(NumberPrototypeToFixed.$call(seconds, 3), "."));
-      const res = hours !== 0 ? `${hours}:${pad(minutes)}` : minutes;
-      return `${res}:${pad(seconds)}.${ms} (${hours !== 0 ? "h:m" : ""}m:ss.mmm)`;
-    }
-
-    if (seconds !== 0) {
-      return `${NumberPrototypeToFixed.$call(seconds, 3)}s`;
-    }
-
-    return `${Number(NumberPrototypeToFixed.$call(ms, 3))}ms`;
-  }
+  const { formatTime } = require("internal/util/debuglog");
 
   const keyKey = "Key";
   const valuesKey = "Values";
