+++ src/js/internal-for-testing.ts
@@ -230,6 +230,7 @@ export const Dequeue = require("internal/fifo");
 // vendored tests need more internals.
 export const exposedInternals = {
   "internal/streams/add-abort-signal": require("internal/streams/add-abort-signal"),
+  "internal/util/debuglog": require("internal/util/debuglog"),
   "internal/async_context_frame": require("internal/async_context_frame"),
   "internal/async_hooks": require("internal/async_hooks"),
   "internal/webstreams/adapters": require("internal/webstreams_adapters"),
